requirements.txt
requests
boto3
pyjwt